+++ src/main/java/com/ds360/komp/model/Warehouse.java
@@ -17,7 +17,7 @@ public class Warehouse {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "warehouse_id")
-    private int warehouseId;
+    private Long warehouseId;
     @Basic
     @Column(name = "title")
     private String title;
